Moved the tests for preferProjectModules into a separate class
